requirements.txt
pendulum==2.1.2
flask-session
sqlalchemy<2.0
requests
pandas
psycopg2-binary